ntomb_mcp/requirements.txt
# ntomb-os-intel MCP Server (Security Analyst Edition) dependencies
mcp>=1.0.0
psutil>=5.9.0
pyyaml>=6.0.0
